requirements.txt
#chromadb==0.4.22
#sentence-transformers==2.2.2
#numpy==1.24.3
#torch==2.1.0
opencv-python>=4.5.0
moviepy>=1.0.3

streamlit>=1.28.0
openai-whisper>=20231117
torch>=2.0.0
librosa>=0.10.0
pydub>=0.25.1
numpy>=1.24.0
pandas>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0